+++ Editor/ButtonDecorator.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Reflection;
 using UnityEditor;
 using UnityEditor.UIElements;
 using UnityEngine;
@@ -64,6 +65,7 @@ namespace Vertx.Attributes.Editor
 	[CustomPropertyDrawer(typeof(ButtonAttribute))]
 	public sealed class ButtonDecorator : DecoratorDrawer
 	{
+#if UNITY_2022_1_OR_NEWER
 		public override VisualElement CreatePropertyGUI()
 		{
 			var buttonAttribute = (ButtonAttribute)attribute;
@@ -75,17 +77,93 @@ namespace Vertx.Attributes.Editor
 		}
 
 		private static Button CreateButton(ButtonAttribute buttonAttribute, SerializedProperty property) =>
-			new(() => Invoke(buttonAttribute, property))
+			new Button(() => Invoke(buttonAttribute, property))
 			{
 				text = string.IsNullOrEmpty(buttonAttribute.DisplayNameOverride)
 					? ObjectNames.NicifyVariableName(buttonAttribute.MethodName)
 					: buttonAttribute.DisplayNameOverride
 			};
 
-		private static void Invoke(ButtonAttribute attribute, SerializedProperty serializedProperty)
+		private static void Invoke(ButtonAttribute attribute, SerializedProperty property)
 		{
-			// TODO write something that invokes the correct method.
+			UnityEngine.Object o = property.serializedObject?.targetObject;
+			
+			if (attribute.StaticMethodType != null)
+			{
+				if (InvokeStatic(attribute, attribute.StaticMethodType, property))
+					return;
+			}
+			
+			if (o == null)
+			{
+				Debug.LogError($"No object was found and {nameof(ButtonAttribute.StaticMethodType)} was not provided to {nameof(ButtonAttribute)}, so no static method can be invoked.");
+				return;
+			}
+
+			Type type = o.GetType();
+			MethodInfo method = type.GetMethod(attribute.MethodName, BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
+			if (method == null)
+			{
+				if (!InvokeStatic(attribute, type, property))
+					Debug.LogError($"Method \"{attribute.MethodName}\" could not be found on {type}.");
+				return;
+			}
+			
+			switch (ValidateParameters(method))
+			{
+				case ParameterResult.None:
+					method.Invoke(o, null);
+					break;
+				case ParameterResult.SerializedProperty:
+					method.Invoke(o, new object[] { property });
+					break;
+				case ParameterResult.Invalid:
+					Debug.LogError($"Parameters on method \"{method}\" must be none, or one {nameof(SerializedProperty)}.");
+					return;
+				default:
+					throw new ArgumentOutOfRangeException();
+			}
 		}
+
+		private static bool InvokeStatic(ButtonAttribute attribute, Type type, SerializedProperty property)
+		{
+			MethodInfo method = type.GetMethod(attribute.MethodName, BindingFlags.Static | BindingFlags.Public | BindingFlags.NonPublic);
+			if (method == null)
+				return false;
+			
+			switch (ValidateParameters(method))
+			{
+				case ParameterResult.None:
+					method.Invoke(null, null);
+					return true;
+				case ParameterResult.SerializedProperty:
+					method.Invoke(null, new object[] { property });
+					return true; 
+				case ParameterResult.Invalid:
+					Debug.LogError($"Parameters on method \"{method}\" must be none, or one {nameof(SerializedProperty)}.");
+					return true;
+				default:
+					throw new ArgumentOutOfRangeException();
+			}
+		}
+
+		enum ParameterResult
+		{
+			None,
+			SerializedProperty,
+			Invalid
+		}
+
+		private static ParameterResult ValidateParameters(MethodInfo method)
+		{
+			ParameterInfo[] parameters = method.GetParameters();
+			if (parameters.Length == 0)
+				return ParameterResult.None;
+			if (parameters.Length != 1)
+				return ParameterResult.Invalid;
+			return parameters[0].ParameterType == typeof(SerializedProperty) ? ParameterResult.SerializedProperty : ParameterResult.Invalid;
+		}
+#endif
 
 		public override void OnGUI(Rect position) => EditorGUI.HelpBox(position, $"{nameof(ButtonAttribute)} is unsupported when using IMGUI.", MessageType.Warning);
 	}