+++ Runtime/ButtonAttribute.cs
@@ -1,10 +1,12 @@
-﻿using UnityEngine;
+﻿using System;
+using UnityEngine;
 
 namespace Vertx.Attributes
 {
 	public sealed class ButtonAttribute : PropertyAttribute
 	{
 		public string MethodName { get; }
+		public Type StaticMethodType { get; }
 
 		public enum Location : byte
 		{
@@ -19,6 +21,12 @@ namespace Vertx.Attributes
 		public ButtonAttribute(string methodName)
 		{
 			MethodName = methodName;
+		}
+		
+		public ButtonAttribute(string methodName, Type staticMethodType)
+		{
+			MethodName = methodName;
+			StaticMethodType = staticMethodType;
 		}
 	}
 }